🐛 FIX: Menu Slider - Fixed menu slider styles
