+++ src/assets/styles/IndexImports.styl
@@ -5,8 +5,6 @@
 @import './components/ButtonContact.styl'
 @import './components/ButtonReel.styl'
 @import './components/Home.styl'
-@import './components/MenuNav.styl'
-@import './components/MenuSlide.styl'
 @import './components/SocialMedia.styl'
 @import './components/Title.styl'
