--- src/assets/styles/components/MenuSlide.styl
@@ -1,30 +1,0 @@
-.MenuSlide__container
-  width 100vw
-  position fixed
-  display flex
-  flex-direction column
-  justify-content flex-start
-  align-items flex-start
-  background-color color8
-  transition 1s ease
-  z-index 20
-  box-shadow 1px 0 4px color7
-
-.MenuSlide__items--container
-  margin 2rem
-  display flex
-  flex-direction column
-  gap 4rem
-  list-style none
-
-.MenuSlide__container.isDark
-  background color5
-  box-shadow none
-
-.MenuSlide__close
-  margin 1.5rem 2rem 0 0
-  align-self flex-end
-  font-size 4.8rem
-  font-weight 400
-  color color3
-
